Se avanzó con la TMDB hasta listar los usurios y mostar sus favoritos

+++ api/routes/favorites.js
@@ -3,13 +3,22 @@ const { Users, Favorites } = require("../models");
 const { validateUser } = require("../middleware/validateUser");
 const router = express.Router();
 
-//Ruta que devuelve todos las movies o series favoritas de un usuario
+//Ruta que devuelve todos las movies o series favoritas del usuario logeado
 router.get("/:category", validateUser, (req, res) => {
   const user = req.user;
   const category = req.params.category;
 
   Favorites.findAll({ where: { userId: user.id, category: category } }).then(
     (movies) => res.status(200).send(movies)
+  );
+});
+
+//Ruta que devuelve todos las movies o series favoritas de un usuario cualquiere
+router.get("/:userId/:category", validateUser, (req, res) => {
+  const { userId, category } = req.params;
+
+  Favorites.findAll({ where: { userId, category } }).then((movies) =>
+    res.status(200).send(movies)
   );
 });
 
@@ -36,9 +45,11 @@ router.delete("/:category/:id", validateUser, (req, res) => {
     .then((movie) => {
       const movieToDelete = movie.dataValues;
       console.log("movieToDelete", movieToDelete);
-      Favorites.destroy({ where: { contentId: id, category } }).then(() => {
-        res.status(200).send(movieToDelete);
-      });
+      Favorites.destroy({ where: { contentId: id, category, userId } }).then(
+        () => {
+          res.status(200).send(movieToDelete);
+        }
+      );
     })
     .catch((error) => res.status(500).send(error));
 });
